+++ client/components/EventList/DurationDisplay.tsx
@@ -1,14 +1,14 @@
-import React from 'react'
+import React, { HTMLProps, FunctionComponent } from 'react'
 import { useTranslation } from 'react-i18next'
 import DateUtils from 'DateUtils'
 import { format } from 'office-ui-fabric'
 
-export interface IDurationDisplayProps extends React.HTMLProps<HTMLDivElement> {
+export interface IDurationDisplayProps extends HTMLProps<HTMLDivElement> {
   displayFormat?: string
   duration: number
 }
 
-export const DurationDisplay: React.FunctionComponent<IDurationDisplayProps> = (
+export const DurationDisplay: FunctionComponent<IDurationDisplayProps> = (
   props: IDurationDisplayProps
 ): JSX.Element => {
   const { t } = useTranslation()