+++ client/pages/Customers/CustomerForm/index.tsx
@@ -35,7 +35,7 @@ export const CustomerForm: FunctionComponent<ICustomerFormProps> = (props: ICust
       }
     })
     if (data?.result.success) {
-      if (props.panel) setTimeout(() => props.panel.onDismiss(null), 1000)
+      if (props.panel) setTimeout(props.panel.onSave, 1000)
       else {
         setMessage({
           text: t('customers.createSuccess', { name: state.model.name }),